src/backend/requirements.txt
celery===4.4.6
coreapi
coreschema
Django
django-cors-headers
django-debug-toolbar
python-dotenv
djangorestframework
djangorestframework-simplejwt
drf-yasg
gunicorn
psycopg2-binary
PyJWT
redis
simplejwt
whitenoise
Pillow
django-filter
